Minor edits and removed adambot-logs

+++ cogs/demographics.py
@@ -1,6 +1,6 @@
 ﻿import discord
 from discord.ext import commands
-from .utils import Permissions, Todo, send_file
+from .utils import Todo, send_file
 import asyncio
 from datetime import datetime
 import matplotlib.pyplot as plt
@@ -8,8 +8,7 @@ from matplotlib.dates import DateFormatter
 
 
 class Demographics(commands.Cog):
-    # Tracks the change in specific roles.
-    # This works over multiple guilds, but cannot function like that due to utils.Permissions.
+    """Tracks the change in specific roles"""
     def __init__(self, bot):
         self.bot = bot
 
@@ -61,20 +60,26 @@ class Demographics(commands.Cog):
             return
 
     @demographics.command(pass_context=True)
-    @commands.has_any_role(*Permissions.STAFF)
     @commands.guild_only()
     async def viewroles(self, ctx):
         """Gets all the roles that are tracked in a guild."""
+        if not await self.bot.is_staff(ctx):
+            await ctx.send("You need the staff role to do this :sob:")
+            return
+
         role_ids = await self._get_roles(ctx.guild)
         roles = [ctx.guild.get_role(x).name for x in role_ids]
         await ctx.send("Currently tracked roles are: " + ', '.join(roles))
 
     @demographics.command(pass_context=True)
-    @commands.has_any_role(*Permissions.STAFF)
     @commands.guild_only()
     async def addrole(self, ctx, role: discord.Role, sample_rate: int = 1):
         """Adds a role to the server's demographic samples.
         `sample_rate` shows how many days are inbetween each sample, and by default is 1."""
+        if not await self.bot.is_staff(ctx):
+            await ctx.send("You need the staff role to do this :sob:")
+            return
+
         def check(m):
             return m.author == ctx.author and m.channel == ctx.channel
 
@@ -99,10 +104,13 @@ class Demographics(commands.Cog):
             question.edit("Unknown response, please try again. :sob:")
 
     @demographics.command(pass_context=True)
-    @commands.has_any_role(*Permissions.STAFF)
     @commands.guild_only()
     async def removerole(self, ctx, role: discord.Role):
         """Gets all the roles that are tracked in a guild."""
+        if not await self.bot.is_staff(ctx):
+            await ctx.send("You need the staff role to do this :sob:")
+            return
+
         def check(m):
             return m.author == ctx.author and m.channel == ctx.channel
 
@@ -125,10 +133,13 @@ class Demographics(commands.Cog):
             await question.edit(content="No action taken.")
 
     @demographics.command(pass_context=True)
-    @commands.has_any_role(*Permissions.STAFF)
     @commands.guild_only()
     async def takesample(self, ctx, role: discord.Role = None):
         """Adds a TODO saying that a sample is required ASAP. If `role` == None then all guild demographics are sampled."""
+        if not await self.bot.is_staff(ctx):
+            await ctx.send("You need the staff role to do this :sob:")
+            return
+
         guild_tracked_roles = await self._get_roles(ctx.guild)
         if not role:
             # Take a sample of all roles
@@ -160,12 +171,15 @@ class Demographics(commands.Cog):
         await ctx.send("A sample has been taken, it may take a few seconds to be registered in the database. :ok_hand:")
 
     @demographics.command(pass_context=True)
-    @commands.has_any_role(*Permissions.STAFF)
     @commands.guild_only()
     async def removeallsamples(self, ctx):
         """Removes all samples from the `demographic_samples` table."""
+        if not await self.bot.is_staff(ctx):
+            await ctx.send("You need the staff role to do this :sob:")
+            return
+
         async with self.bot.pool.acquire() as connection:
-            await connection.execute("DELETE FROM demographic_samples;")
+            await connection.execute("DELETE FROM demographic_samples WHERE role_reference IN (SELECT id FROM demographic_roles WHERE guild_id = $1);", ctx.guild.id) # Removes samples for that guild
         await ctx.send("All samples have been deleted. :sob:")
 
     # Error handlers
@@ -183,7 +197,6 @@ class Demographics(commands.Cog):
 
     # Old command
     @demographics.command(pass_context=True)
-    @commands.has_any_role(*Permissions.MEMBERS)
     @commands.guild_only()
     async def show(self, ctx):
         """View server demographics."""  # DO NOT REMOVE THIS METHOD (if you plan on removing, remove dependency in demographics command group declaration)
@@ -200,7 +213,6 @@ class Demographics(commands.Cog):
         await ctx.send(message)
 
     @demographics.command(pass_context=True)
-    @commands.has_any_role(*Permissions.MEMBERS)
     @commands.guild_only()
     async def chart(self, ctx):
         """View a guild's demographics over time"""
